backend/requirements.txt
Flask==2.0.1
Flask-PyMongo==2.3.0
pymongo==3.12.0
python-dotenv==0.19.0
Werkzeug==2.0.1